+++ ios/vendored/unversioned/react-native-reanimated/apple/LayoutReanimation/REASwizzledUIManager.h
@@ -1,0 +1,6 @@
+#import <RNReanimated/REAAnimationsManager.h>
+
+@interface REASwizzledUIManager : NSObject
+- (instancetype)initWithUIManager:(RCTUIManager *)uiManager
+             withAnimationManager:(REAAnimationsManager *)animationsManager;
+@end