+++ packages/expo-image/ios/ImageModule.swift
@@ -70,7 +70,9 @@ public final class ImageModule: Module {
       }
 
       Prop("enableLiveTextInteraction") { (view, enableLiveTextInteraction: Bool?) in
+        #if !os(tvOS)
         view.enableLiveTextInteraction = enableLiveTextInteraction ?? false
+        #endif
       }
 
       Prop("accessible") { (view, accessible: Bool?) in
@@ -93,6 +95,18 @@ public final class ImageModule: Module {
         view.intrinsicSizes = intrinsicSize
       }
 
+      Prop("autoplay") { (view, autoplay: Bool?) in
+        view.sdImageView.autoPlayAnimatedImage = autoplay ?? true
+      }
+
+      AsyncFunction("startAnimating") { (view: ImageView) in
+        view.sdImageView.startAnimating()
+      }
+
+      AsyncFunction("stopAnimating") { (view: ImageView) in
+        view.sdImageView.stopAnimating()
+      }
+
       OnViewDidUpdateProps { view in
         view.onPropsUpdated()
       }
@@ -112,10 +126,26 @@ public final class ImageModule: Module {
         promise.resolve(true)
       }
     }
+
+    AsyncFunction("getCachePathAsync") { (cacheKey: String, promise: Promise) in
+      /*
+       We need to check if the image exists in the cache first since `cachePath` will
+       return a path regardless of whether or not the image exists.
+       */
+      SDImageCache.shared.diskImageExists(withKey: cacheKey) { exists in
+        if exists {
+          let cachePath = SDImageCache.shared.cachePath(forKey: cacheKey)
+
+          promise.resolve(cachePath)
+        } else {
+          promise.resolve(nil)
+        }
+      }
+    }
   }
 
   static func registerCoders() {
-    if #available(iOS 14.0, *) {
+    if #available(iOS 14.0, tvOS 14.0, *) {
       // By default Animated WebP is not supported
       SDImageCodersManager.shared.addCoder(SDImageAWebPCoder.shared)
     } else {