+++ packages/expo-image/src/Image.types.ts
@@ -1,5 +1,7 @@
 import { ImageStyle as RNImageStyle, ViewProps, StyleProp, ViewStyle } from 'react-native';
 
+import ExpoImage from './ExpoImage';
+
 export type ImageSource = {
   /**
    * A string representing the resource identifier for the image,
@@ -200,6 +202,14 @@ export interface ImageProps extends ViewProps {
   recyclingKey?: string | null;
 
   /**
+   * Determines if an image should automatically begin playing if it is an
+   * animated image.
+   * @default true
+   * @platform ios
+   */
+  autoplay?: boolean;
+
+  /**
    * Called when the image starts to load.
    */
   onLoadStart?: () => void;
@@ -324,6 +334,8 @@ export interface ImageNativeProps extends ImageProps {
   placeholder?: ImageSource[];
   contentPosition?: ImageContentPositionObject;
   transition?: ImageTransition | null;
+  autoplay?: boolean;
+  nativeViewRef?: React.RefObject<ExpoImage>;
 }
 
 /**
@@ -458,6 +470,7 @@ export type ImageLoadEventData = {
     width: number;
     height: number;
     mediaType: string | null;
+    isAnimated?: boolean;
   };
 };