+++ packages/expo-updates/ios/EXUpdates/UpdatesModule.swift
@@ -15,12 +15,6 @@ import ExpoModulesCore
  * by EXUpdatesBinding, a scoped module, in Expo Go.
  */
 public final class UpdatesModule: Module {
-  private let methodQueue = UpdatesUtils.methodQueue
-
-  public required init(appContext: AppContext) {
-    super.init(appContext: appContext)
-  }
-
   // swiftlint:disable cyclomatic_complexity
   public func definition() -> ModuleDefinition {
     Name("ExpoUpdates")
@@ -46,8 +40,7 @@ public final class UpdatesModule: Module {
         ]
       }
 
-      let database = AppController.sharedInstance.database
-      let embeddedUpdate = EmbeddedAppLoader.embeddedManifest(withConfig: config, database: database)
+      let embeddedUpdate = AppController.sharedInstance.getEmbeddedUpdate()
       let isEmbeddedLaunch = embeddedUpdate != nil && embeddedUpdate?.updateId == launchedUpdate.updateId
 
       let commitTime = UInt64(floor(launchedUpdate.commitTime.timeIntervalSince1970 * 1000))
@@ -87,42 +80,40 @@ public final class UpdatesModule: Module {
     }
 
     AsyncFunction("checkForUpdateAsync") { (promise: Promise) in
-      UpdatesUtils.checkForUpdate { result in
-        if result["message"] != nil {
-          guard let message = result["message"] as? String else {
-            promise.reject("ERR_UPDATES_CHECK", "")
-            return
-          }
-          promise.reject("ERR_UPDATES_CHECK", message)
-          return
-        }
-        if result["manifest"] != nil {
-          promise.resolve([
-            "isAvailable": true,
-            "manifest": result["manifest"],
-            "isRollBackToEmbedded": false
-          ])
-          return
-        }
-        if result["isRollBackToEmbedded"] != nil {
-          promise.resolve([
-            "isAvailable": false,
-            "isRollBackToEmbedded": result["isRollBackToEmbedded"]
-          ])
-          return
-        }
-        if result["reason"] != nil {
+      let config = AppController.sharedInstance.config
+      guard config.isEnabled else {
+        throw UpdatesDisabledException()
+      }
+      guard AppController.sharedInstance.isStarted else {
+        throw UpdatesNotInitializedException()
+      }
+
+      AppController.sharedInstance.checkForUpdate { remoteCheckResult in
+        switch remoteCheckResult {
+        case .noUpdateAvailable(let reason):
           promise.resolve([
             "isAvailable": false,
             "isRollBackToEmbedded": false,
-            "reason": result["reason"]
+            "reason": reason
           ])
           return
+        case .updateAvailable(let manifest):
+          promise.resolve([
+            "isAvailable": true,
+            "manifest": manifest,
+            "isRollBackToEmbedded": false
+          ])
+          return
+        case .rollBackToEmbedded:
+          promise.resolve([
+            "isAvailable": false,
+            "isRollBackToEmbedded": true
+          ])
+          return
+        case .error(let error):
+          promise.reject("ERR_UPDATES_CHECK", error.localizedDescription)
+          return
         }
-        promise.resolve([
-          "isAvailable": false,
-          "isRollBackToEmbedded": false
-        ])
       }
     }
 
@@ -131,17 +122,14 @@ public final class UpdatesModule: Module {
       guard config.isEnabled else {
         throw UpdatesDisabledException()
       }
-
-      guard let scopeKey = config.scopeKey else {
-        throw Exception(name: "ERR_UPDATES_SCOPE_KEY", description: "Muse have scopeKey in config")
+      guard AppController.sharedInstance.isStarted else {
+        throw UpdatesNotInitializedException()
       }
 
-      AppController.sharedInstance.database.databaseQueue.async {
-        do {
-          promise.resolve(try AppController.sharedInstance.database.extraParams(withScopeKey: scopeKey))
-        } catch {
-          promise.reject(error)
-        }
+      AppController.sharedInstance.getExtraParams { extraParams in
+        promise.resolve(extraParams)
+      } error: { error in
+        promise.reject(error)
       }
     }
 
@@ -150,18 +138,14 @@ public final class UpdatesModule: Module {
       guard config.isEnabled else {
         throw UpdatesDisabledException()
       }
-
-      guard let scopeKey = config.scopeKey else {
-        throw Exception(name: "ERR_UPDATES_SCOPE_KEY", description: "Muse have scopeKey in config")
+      guard AppController.sharedInstance.isStarted else {
+        throw UpdatesNotInitializedException()
       }
 
-      AppController.sharedInstance.database.databaseQueue.async {
-        do {
-          try AppController.sharedInstance.database.setExtraParam(key: key, value: value, withScopeKey: scopeKey)
-          promise.resolve(nil)
-        } catch {
-          promise.reject(error)
-        }
+      AppController.sharedInstance.setExtraParam(key: key, value: value) {
+        promise.resolve(nil)
+      } error: { error in
+        promise.reject(error)
       }
     }
 
@@ -185,16 +169,38 @@ public final class UpdatesModule: Module {
     }
 
     AsyncFunction("fetchUpdateAsync") { (promise: Promise) in
-      UpdatesUtils.fetchUpdate { result in
-        if result["message"] != nil {
-          guard let message = result["message"] as? String else {
-            promise.reject("ERR_UPDATES_FETCH", "")
-            return
-          }
-          promise.reject("ERR_UPDATES_FETCH", message)
+      let config = AppController.sharedInstance.config
+      guard config.isEnabled else {
+        throw UpdatesDisabledException()
+      }
+      guard AppController.sharedInstance.isStarted else {
+        throw UpdatesNotInitializedException()
+      }
+
+      AppController.sharedInstance.fetchUpdate { fetchUpdateResult in
+        switch fetchUpdateResult {
+        case .success(let manifest):
+          promise.resolve([
+            "isNew": true,
+            "isRollBackToEmbedded": false,
+            "manifest": manifest
+          ])
           return
-        } else {
-          promise.resolve(result)
+        case .failure:
+          promise.resolve([
+            "isNew": false,
+            "isRollBackToEmbedded": false
+          ])
+          return
+        case .rollBackToEmbedded:
+          promise.resolve([
+            "isNew": false,
+            "isRollBackToEmbedded": true
+          ])
+          return
+        case .error(let error):
+          promise.reject("ERR_UPDATES_FETCH", error.localizedDescription)
+          return
         }
       }
     }
@@ -202,17 +208,16 @@ public final class UpdatesModule: Module {
     // Getter used internally by useUpdates()
     // to initialize its state
     AsyncFunction("getNativeStateMachineContextAsync") { (promise: Promise) in
-      UpdatesUtils.getNativeStateMachineContextJson { result in
-        if result["message"] != nil {
-          guard let message = result["message"] as? String else {
-            promise.reject("ERR_UPDATES_CHECK", "")
-            return
-          }
-          promise.reject("ERR_UPDATES_CHECK", message)
-          return
-        } else {
-          promise.resolve(result)
-        }
+      let config = AppController.sharedInstance.config
+      guard config.isEnabled else {
+        throw UpdatesDisabledException()
+      }
+      guard AppController.sharedInstance.isStarted else {
+        throw UpdatesNotInitializedException()
+      }
+
+      AppController.sharedInstance.getNativeStateMachineContext { stateMachineContext in
+        promise.resolve(stateMachineContext.json)
       }
     }
   }